+++ src/clay_board_style_system/core/computed_style.nim
@@ -8,7 +8,9 @@ type
 
   FlexDirection* = enum
     fdRow,
-    fdColumn
+    fdColumn,
+    fdRowReverse,
+    fdColumnReverse
 
   FlexWrap* = enum
     fwNoWrap,